scripts/requirements.txt
pandas==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
openpyxl==3.1.2 